# Conflicts:
#	api/urls.py
#	l2-frontend/src/fields/RadioFieldById.vue
#	l2-frontend/src/mainWithRouter.ts

+++ results/views.py
@@ -45,7 +45,7 @@ import slog.models as slog
 from api.stationar.stationar_func import hosp_get_hosp_direction
 from appconf.manager import SettingManager
 from clients.models import CardBase
-from directions.models import Issledovaniya, Result, Napravleniya, ParaclinicResult, Recipe, DirectionDocument, DocumentSign
+from directions.models import Issledovaniya, Result, Napravleniya, ParaclinicResult, Recipe, DirectionDocument, DocumentSign, IssledovaniyaFiles, ComplexResearchAccountPerson
 from laboratory.decorators import logged_in_or_token
 from laboratory.settings import (
     DEATH_RESEARCH_PK,
@@ -57,6 +57,7 @@ from laboratory.settings import (
     RESEARCHES_NOT_PRINT_FOOTERS,
     RESULT_LABORATORY_FORM,
     SELF_WATERMARKS,
+    DISABLE_PATIENT_CANVAS_MARKER,
 )
 from laboratory.settings import FONTS_FOLDER
 from laboratory.utils import strdate
@@ -131,7 +132,6 @@ def result_print(request):
     interactive_text_field = SettingManager.get("interactive_text_field", default='False', default_type='b')
 
     buffer = BytesIO()
-
     split = request.GET.get("split", "1") == "1"
     protocol_plain_text = request.GET.get("protocol_plain_text", "0") == "1"
     leftnone = request.GET.get("leftnone", "0") == "0"
@@ -140,7 +140,11 @@ def result_print(request):
     if med_certificate:
         med_certificate_title = "Справка - "
     hosp = request.GET.get("hosp", "0") == "1"
+    complex = request.GET.get("complex", "0") == "1"
     with_signature_stamps = request.GET.get("withSignatureStamps", "0") == "1"
+
+    if complex:
+        pk = ComplexResearchAccountPerson.get_complex_confirm_directions(tuple(pk))
 
     doc = BaseDocTemplate(
         buffer,
@@ -365,7 +369,7 @@ def result_print(request):
             canvas_mark.drawString(155 * mm, 285 * mm, '{}'.format(" НЕ ПОДТВЕРЖДЕНО "))
             canvas_mark.setFont('FreeSans', 12)
             canvas_mark.drawString(175 * mm, 281 * mm, '{}'.format("( образец )"))
-        if not watermarks and not DEATH_RESEARCH_PK and not GISTOLOGY_RESEARCH_PK and not SELF_WATERMARKS:
+        if not watermarks and not DEATH_RESEARCH_PK and not GISTOLOGY_RESEARCH_PK and not SELF_WATERMARKS and not DISABLE_PATIENT_CANVAS_MARKER:
             if direction.hospital:
                 canvas_mark.drawString(55 * mm, 13 * mm, direction.hospital.safe_short_title)
             else:
@@ -376,8 +380,9 @@ def result_print(request):
                 canvas_mark.drawString(
                     55 * mm, 9.6 * mm, '№ карты: {}; Номер: {} {}; Направление № {}'.format(direction.client.number_with_type(), num_card, number_poliklinika, direction.pk)
                 )
-            canvas_mark.drawString(55 * mm, 7.1 * mm, 'Пациент: {} {}'.format(direction.client.individual.fio(), individual_birthday))
-            canvas_mark.line(55 * mm, 12.7 * mm, 181 * mm, 11.5 * mm)
+            if not DISABLE_PATIENT_CANVAS_MARKER:
+                canvas_mark.drawString(55 * mm, 7.1 * mm, 'Пациент: {} {}'.format(direction.client.individual.fio(), individual_birthday))
+                canvas_mark.line(55 * mm, 12.7 * mm, 181 * mm, 11.5 * mm)
             if qr_data:
                 qr_code = qr.QrCodeWidget(qr_data)
                 qr_code.barWidth = 15 * mm
@@ -451,6 +456,10 @@ def result_print(request):
                 has_paraclinic = True
             if iss.link_file:
                 link_result.append(iss.link_file)
+                link_files = True
+            if IssledovaniyaFiles.objects.filter(issledovaniye=iss).first():
+                iss_uploaded_file = IssledovaniyaFiles.objects.filter(issledovaniye=iss).first()
+                link_result.append(iss_uploaded_file.uploaded_file.path)
                 link_files = True
             if 'выпис' in iss.research.title.lower():
                 is_extract = True
@@ -793,12 +802,15 @@ def result_print(request):
         date_now1 = datetime.datetime.strftime(datetime.datetime.now(), "%y%m%d%H%M%S")
         date_now_str = str(random.random()) + str(date_now1)
         dir_param = SettingManager.get("dir_param", default='/tmp', default_type='s')
-        file_dir_l2 = os.path.join(dir_param, date_now_str + '_dir.pdf')
+
         buffer.seek(0)
-        save(buffer, filename=file_dir_l2)
-        dst_dir = SettingManager.get("root_dir")
-        file_dir = [os.path.join(dst_dir, link_f) for link_f in link_result]
-        file_dir.append(file_dir_l2)
+        file_dir_l2 = None
+        if buffer.getbuffer().nbytes > 0:
+            file_dir_l2 = os.path.join(dir_param, date_now_str + '_dir.pdf')
+            save(buffer, filename=file_dir_l2)
+        file_dir = [link_f for link_f in link_result]
+        if file_dir_l2:
+            file_dir.append(file_dir_l2)
         writer = PdfWriter()
         pdf_all = BytesIO()
         for inpfn in file_dir:
@@ -808,7 +820,8 @@ def result_print(request):
         pdf_all.close()
         response.write(pdf_out)
         buffer.close()
-        os.remove(file_dir_l2)
+        if file_dir_l2:
+            os.remove(file_dir_l2)
         return response
 
     pdf = buffer.getvalue()
